services/ingestor/pyproject.toml
[project]
name = "ingestor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.37.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
